Add checkConnection function
The checkConnection function is added. It will check the connection to the server every 3 seconds, which time interval can be adjusted.

You may see "Not connected" at the upper right corner of the website and a warning message pops up in the console every 3 seconds due to the request limit of the mock server. Trying to solve it.